# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FonAd / DAADD API
  version: 1.0.0
  description: |
    FonAd (DAADD) is a full-featured digital advertising and analytics platform.
    It provides campaign management, real-time event tracking, AI-driven optimization,
    anomaly detection, cross-device attribution, benchmarking, storytelling reports,
    viewer reward programs, team collaboration, and notification services.
  contact:
    name: FonAd Engineering
    email: engineering@fonad.io
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      description: Returns a paginated, filterable list of campaigns with impression and click counts.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CampaignStatus'
        - name: industry
          in: query
          schema:
            $ref: '#/components/schemas/CampaignIndustry'
        - name: advertiser_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by advertiser UUID
        - name: search
          in: query
          schema:
            type: string
          description: Case-insensitive name search
      responses:
        '200':
          description: Paginated list of campaigns
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Campaign'
                            - type: object
                              properties:
                                impressions:
                                  type: integer
                                clicks:
                                  type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Campaigns]
      summary: Create a campaign
      description: |
        Creates a new campaign in DRAFT status. Requires advertiser, admin, or campaign_manager role.
        Budget must be at least $500 for campaigns with regional targeting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /campaigns/{id}:
    get:
      tags: [Campaigns]
      summary: Get campaign by ID
      description: Returns a single campaign with advertiser and creatives relations.
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: Campaign detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Campaign'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Campaigns]
      summary: Update a campaign
      description: Updates campaign fields. Only the owner or team members with editor/admin role can update.
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Campaign updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Campaigns]
      summary: Delete a campaign
      description: |
        Deletes a DRAFT campaign permanently. Non-draft campaigns are set to COMPLETED status instead.
        Only the campaign owner can delete.
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: Campaign deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /campaigns/{id}/toggle-ai:
    patch:
      tags: [Campaigns]
      summary: Toggle AI optimization
      description: |
        Enables or disables AI optimization for a campaign.
        AI can only be enabled after 48 hours of campaign runtime.
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                  description: Whether to enable or disable AI optimization
                mode:
                  $ref: '#/components/schemas/AiMode'
      responses:
        '200':
          description: AI optimization updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /campaigns/{id}/creatives:
    post:
      tags: [Campaigns]
      summary: Upload a creative asset
      description: Uploads an image or video creative asset for a campaign.
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image (png, jpg, gif, webp) or video (mp4, webm, mov) file
      responses:
        '201':
          description: Creative uploaded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Creative'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────── Analytics ───────────────────────────
...
  /events/campaign/{campaignId}:
    get:
      tags: [Events]
      summary: Get events for a campaign
      description: Returns a paginated and filterable list of events for a given campaign.
      parameters:
        - $ref: '#/components/parameters/CampaignIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - name: event_type
          in: query
          schema:
            $ref: '#/components/schemas/EventType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - name: device_type
          in: query
          schema:
            $ref: '#/components/schemas/DeviceType'
      responses:
        '200':
          description: Paginated list of events
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────── Heatmaps ───────────────────────────
...
  /teams/campaign/{campaignId}:
    get:
      tags: [Teams]
      summary: List team members
      description: Returns all team members for a campaign.
      parameters:
        - $ref: '#/components/parameters/CampaignIdPath'
      responses:
        '200':
          description: List of team members
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamMember'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Sort:
      name: sort
      in: query
      schema:
        type: string
        default: created_at
      description: Field name to sort by
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [ASC, DESC]
        default: DESC
      description: Sort order
    CampaignId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Campaign UUID
    CampaignIdPath:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Campaign UUID
...
    StartDate:
      name: start_date
      in: query
      schema:
        type: string
        format: date-time
      description: Filter start date (ISO 8601)
    EndDate:
      name: end_date
      in: query
      schema:
        type: string
        format: date-time
      description: Filter end date (ISO 8601)
  responses:
    BadRequest:
      description: Bad request -- missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized -- missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden -- insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserRole:
      type: string
      enum: [advertiser, campaign_manager, analyst, end_user, admin]
    CampaignStatus:
      type: string
      enum: [draft, active, paused, completed, suspended]
    CampaignIndustry:
      type: string
      enum:
        - entertainment
        - sports
        - retail
        - technology
        - food_beverage
        - health
        - finance
        - education
        - travel
        - automotive
        - other
    AiMode:
      type: string
      enum: [auto_adjust, recommendation_only]
    EventType:
      type: string
      enum: [impression, click, conversion, share, reward_claim, ad_view, ad_dismiss]
    DeviceType:
      type: string
      enum: [desktop, mobile, tablet]
    CreativeType:
      type: string
      enum: [image, video]
...
    TeamRole:
      type: string
      enum: [viewer, editor, admin]
    InviteStatus:
      type: string
      enum: [pending, accepted, rejected]

    # ─────────────── Entity Schemas ───────────────
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, name, role, is_active, created_at, updated_at]
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        advertiser_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        budget_total:
          type: number
          format: float
          description: Total campaign budget in USD
        budget_spent:
          type: number
          format: float
          description: Amount of budget spent so far
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        industry:
          $ref: '#/components/schemas/CampaignIndustry'
        targeting_config:
          $ref: '#/components/schemas/TargetingConfig'
        status:
          $ref: '#/components/schemas/CampaignStatus'
        ai_optimization_enabled:
          type: boolean
        ai_mode:
          $ref: '#/components/schemas/AiMode'
        reward_value:
          type: number
          format: float
          description: Reward amount per qualifying event
        is_age_restricted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        advertiser:
          $ref: '#/components/schemas/User'
        creatives:
          type: array
          items:
            $ref: '#/components/schemas/Creative'
      required:
        - id
        - advertiser_id
        - name
        - budget_total
        - budget_spent
        - start_date
        - end_date
        - industry
        - status
        - ai_optimization_enabled
        - ai_mode
        - reward_value
        - is_age_restricted
        - created_at
        - updated_at
    TargetingConfig:
      type: object
      nullable: true
      properties:
        age_range:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        regions:
          type: array
          items:
            type: string
        devices:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
    Creative:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/CreativeType'
        url:
          type: string
          description: Relative URL path to the uploaded file
        format:
          type: string
          description: File extension (e.g. png, mp4)
        file_size:
          type: integer
          description: File size in bytes
        original_name:
          type: string
          description: Original uploaded file name
        created_at:
          type: string
          format: date-time
      required: [id, campaign_id, type, url, format, file_size, original_name, created_at]
    AdEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_type:
          $ref: '#/components/schemas/EventType'
        campaign_id:
          type: string
          format: uuid
        ad_id:
          type: string
          nullable: true
        user_id:
          type: string
          format: uuid
          nullable: true
        device_type:
          $ref: '#/components/schemas/DeviceType'
        device_id:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        geo_lat:
          type: number
          format: float
          nullable: true
        geo_lng:
          type: number
          format: float
          nullable: true
        geo_city:
          type: string
          nullable: true
        geo_country:
          type: string
          nullable: true
        unified_user_id:
          type: string
          nullable: true
          description: Cross-device unified user identifier
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, event_type, campaign_id, device_type, created_at]
...
    TeamMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/TeamRole'
        invited_by:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/InviteStatus'
        created_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
      required: [id, user_id, campaign_id, role, invited_by, status, created_at]
...
    CreateCampaignRequest:
      type: object
      required: [name, budget_total, start_date, end_date]
      properties:
        name:
          type: string
        description:
          type: string
        budget_total:
          type: number
          format: float
          description: Total budget in USD. Must be >= $500 if regional targeting is used.
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        industry:
          $ref: '#/components/schemas/CampaignIndustry'
        targeting_config:
          $ref: '#/components/schemas/TargetingConfig'
        reward_value:
          type: number
          format: float
          description: Per-event reward amount
        is_age_restricted:
          type: boolean
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        budget_total:
          type: number
          format: float
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        industry:
          $ref: '#/components/schemas/CampaignIndustry'
        targeting_config:
          $ref: '#/components/schemas/TargetingConfig'
        reward_value:
          type: number
          format: float
        is_age_restricted:
          type: boolean
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Response payload (type varies by endpoint)
      required: [success]
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: Total number of records
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Items per page
            totalPages:
              type: integer
              description: Total number of pages
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
          required: [total, page, limit, totalPages, hasNext, hasPrev]
      required: [success, data, pagination]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_ERROR
                - INVALID_CREDENTIALS
                - EMAIL_EXISTS
                - ACCOUNT_INACTIVE
                - INVALID_TRANSITION
                - INVALID_BUDGET
                - AI_NOT_READY
            details:
              description: Additional error context
          required: [message, code]
      required: [success, error]